license_project/requirements.txt
asgiref==3.9.1
brevo-python==1.2.0
certifi==2025.8.3
cffi==1.17.1
cryptography==45.0.6
dj-database-url==3.0.1
Django==5.2.5
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
